# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "BatchQueryRequest": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchQueryItem"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchQueryRequest",
        "type": "object"
      },
      "BatchQueryResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchResult"
            },
            "title": "Items",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchQueryResponse",
        "type": "object"
      },
...
      "ContextEnvelope": {
        "additionalProperties": false,
        "properties": {
          "comparisons": {
            "items": {
              "$ref": "#/components/schemas/Comparison"
            },
            "title": "Comparisons",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_chunks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence Chunks",
            "type": "array"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "title": "Facts",
            "type": "array"
          },
          "field_availability": {
            "items": {
              "$ref": "#/components/schemas/FieldAvailability"
            },
            "title": "Field Availability",
            "type": "array"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "insight_cards": {
            "items": {
              "$ref": "#/components/schemas/InsightCard"
            },
            "title": "Insight Cards",
            "type": "array"
          },
          "metric_records": {
            "items": {
              "$ref": "#/components/schemas/Metric"
            },
            "title": "Metric Records",
            "type": "array"
          },
          "metrics": {
            "additionalProperties": true,
            "description": "Backward-compatible metric name to value/metadata mapping",
            "title": "Metrics",
            "type": "object"
          },
          "pipeline_state": {
            "items": {
              "$ref": "#/components/schemas/PipelineRun"
            },
            "title": "Pipeline State",
            "type": "array"
          },
          "pipeline_status": {
            "$ref": "#/components/schemas/FilingStatus"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotValues"
          },
          "source_documents": {
            "items": {
              "$ref": "#/components/schemas/SourceDocument"
            },
            "title": "Source Documents",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ReadinessStatus"
          },
          "validation": {
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "title": "Validation",
            "type": "array"
          }
        },
        "required": [
          "identity",
          "status",
          "pipeline_status",
          "freshness",
          "quality",
          "snapshot",
          "metrics",
          "metric_records",
          "events",
          "evidence",
          "facts",
          "field_availability",
          "validation",
          "comparisons",
          "insight_cards",
          "source_documents",
          "pipeline_state"
        ],
        "title": "ContextEnvelope",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FilingStatus": {
        "enum": [
          "pending",
          "ingested",
          "extracted",
          "validated",
          "insight_ready",
          "failed"
        ],
        "title": "FilingStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataState"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "created_at",
          "updated_at",
          "identity"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ReadinessStatus": {
        "enum": [
          "ready",
          "processing",
          "stale"
        ],
        "title": "ReadinessStatus",
        "type": "string"
      },
      "RefreshResponse": {
        "additionalProperties": false,
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status",
          "identity"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FinancialReports API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/batch/filings/query": {
      "post": {
        "operationId": "batch_query_v1_batch_filings_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Query",
        "tags": [
          "batch"
        ]
      }
    },
...
    "/v1/filings/{stock_code}/{period}/context": {
      "get": {
        "operationId": "context_v1_filings__stock_code___period__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fields"
            }
          },
          {
            "in": "query",
            "name": "evidence_limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 0,
              "title": "Evidence Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Context",
        "tags": [
          "filings"
        ]
      }
    },
    "/v1/filings/{stock_code}/{period}/refresh": {
      "post": {
        "operationId": "refresh_v1_filings__stock_code___period__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/filings/{stock_code}/{period}/snapshot": {
      "get": {
        "operationId": "snapshot_v1_filings__stock_code___period__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fields"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Snapshot",
        "tags": [
          "filings"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/stocks": {
      "get": {
        "operationId": "list_stocks_v1_stocks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "industry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StocksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Stocks",
        "tags": [
          "filings"
        ]
      }
    },
    "/v1/stocks/{stock_code}/periods": {
      "get": {
        "operationId": "list_periods_v1_stocks__stock_code__periods_get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "title": "Stock Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Periods",
        "tags": [
          "filings"
        ]
      }
    }
  }
}